+++ example/lib/ui_widget_gallery/imageview_page.dart
@@ -32,7 +32,7 @@ class ImageViewPage extends StatelessWidget {
         Divider().pSymmetric(v: 16),
         '40px'.text.bold.make().pOnly(bottom: 12),
         HStack([
-          ExImageView(url: '', size: 40).pOnly(right: 12),
+          ExImageViewCached(url: '', size: 40).pOnly(right: 12),
           ExImageView(url: '', size: 40).pOnly(right: 12),
           ExImageView(url: 'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300', size: 40).pOnly(right: 12),
           ExImageView(url: 'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300', size: 40).pOnly(right: 12),