+++ lib/src/features/ex_datepicker.dart
@@ -11,6 +11,8 @@ mixin ExDatePicker {
   // DATE PICKER (CUPERTINO) —————————————————————————————————————————————————
   // —————————————————————————————————————————————————————————————————————————
   static void datePickerCupertino({
+    String title = 'Pilih Tanggal',
+    String btnText = 'Simpan',
     DateTime? initialDate,
     DateTime? minDate,
     DateTime? maxDate,
@@ -20,34 +22,41 @@ mixin ExDatePicker {
     var output = DateTime.now();
     showCupertinoModalPopup(
       context: Get.context!,
-      builder: (_) => Container(
-        height: 270,
-        decoration: ExDecorator.boxBottomSheet(),
-        child: Column(
-          children: [
-            HStack([
-              // 'Select Date'.text.size(16).bold.make(),
-              Spacer(),
-              CupertinoButton(
-                child: 'Apply'.text.make(),
+      builder: (_) => CupertinoApp(
+        debugShowCheckedModeBanner: false,
+        builder: (context, child) => Container(
+          height: Get.size.height * 0.40,
+          // color: colorPrimary[300],
+          decoration: ExDecorator.boxBottomSheet(),
+          child: Column(
+            children: [
+              HStack([
+                title.text.size(18).bold.black.make(),
+                Spacer(),
+              ]).pOnly(left: 16, top: 16, bottom: 8),
+              Divider(),
+              CupertinoTheme(
+                data: CupertinoThemeData(brightness: Brightness.light),
+                child: CupertinoDatePicker(
+                  mode: mode,
+                  initialDateTime: initialDate ?? DateTime.now(),
+                  minimumDate: minDate ?? DateTime.now().subtract(1.seconds),
+                  // somehow must like this =,=
+                  maximumDate: maxDate?.add(1.minutes) ?? DateTime(2050),
+                  onDateTimeChanged: (val) => output = val,
+                  dateOrder: DatePickerDateOrder.dmy,
+                ),
+              ).expand(),
+              ExButtonDefault(
+                width: double.infinity,
+                label: btnText,
                 onPressed: () {
                   callback.call(output);
-                  Get.back();
+                  Navigator.of(Get.context!).pop();
                 },
-              ),
-            ]).pSymmetric(h: 16, v: 2),
-            SizedBox(
-              height: 200,
-              child: CupertinoDatePicker(
-                mode: mode,
-                initialDateTime: initialDate ?? DateTime.now(),
-                minimumDate: minDate ?? DateTime.now().subtract(1.seconds),
-                maximumDate: maxDate ?? DateTime(2050),
-                onDateTimeChanged: (val) => output = val,
-                use24hFormat: true,
-              ),
-            ),
-          ],
+              ).pSymmetric(h: 24).pOnly(bottom: 24),
+            ],
+          ),
         ),
       ),
     );