+++ lib/src/widgets/ex_image_no_cache_view.dart
@@ -1,22 +1,39 @@
 import 'dart:typed_data';
 
+import 'package:ex/src/color.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 
 import '../../ex.dart';
 
 class ExImageNoCacheView extends StatelessWidget {
-  const ExImageNoCacheView(
-    this.imageUrl, {
-    this.headers,
+  const ExImageNoCacheView({
     Key? key,
+    required this.url,
+    this.headers,
+    this.height,
+    this.width,
+    this.size,
+    this.borderColor,
+    this.borderWidth,
+    this.radius = 8,
+    this.boxFit = BoxFit.cover,
+    this.errorWidget,
   }) : super(key: key);
 
-  final String imageUrl;
+  final String url;
   final Map<String, String>? headers;
+  final double? height;
+  final double? width;
+  final double? size;
+  final Color? borderColor;
+  final double? borderWidth;
+  final double radius;
+  final BoxFit boxFit;
+  final Widget? errorWidget;
 
   Future<Uint8List> getImageBytes() async {
-    final response = await get(Uri.parse(imageUrl), headers: headers);
+    final response = await get(Uri.parse(url), headers: headers);
     return response.bodyBytes;
   }
 
@@ -36,12 +53,11 @@ class ExImageNoCacheView extends StatelessWidget {
             },
           ).backgroundColor(Colors.white);
         }
-        return Center(
-          child: LoadingAnimationWidget.discreteCircle(
-            color: Colors.grey,
-            size: 32,
-          ),
-        );
+        return SizedBox(
+          width: 24,
+          height: 24,
+          child: Container(color: colorNeutral),
+        ).shimmer(primaryColor: colorNeutral[100]!, secondaryColor: colorNeutral[200]);
       },
     );
   }